Fix CI: replace missing build.cmd with inline zig + msbuild steps; update action versions (#14)
* Fix CI: replace missing build.cmd with inline zig cc + msbuild steps

Agent-Logs-Url: https://github.com/joelvaneenwyk/dwm-windows/sessions/89c634fc-59ef-418d-a5e6-a5d920bda9a9


* Update action versions: checkout@v6, setup-msbuild@v3, action-gh-release@v3, codeql-action@v4

Agent-Logs-Url: https://github.com/joelvaneenwyk/dwm-windows/sessions/c6808701-2886-4d9f-8b6b-aa8447a16088


* Remove legacy VS2008 hidapi .sln files that break NuGet auto-submission

Agent-Logs-Url: https://github.com/joelvaneenwyk/dwm-windows/sessions/f652870b-d268-4f6e-91ef-037e0f744c67


* Fix C4388: cast LENGTH(keys) to int to eliminate signed/unsigned mismatch

Agent-Logs-Url: https://github.com/joelvaneenwyk/dwm-windows/sessions/7cccd927-22f4-49cf-9631-34ca93011a84


---------

+++ src/dwm-win32.c
@@ -133,7 +133,7 @@ cleanup(lua_State* L) {
         KillTimer(barhwnd, 1);
     }
 
-    for (i = 0; i < LENGTH(keys); i++) {
+    for (i = 0; i < (int)LENGTH(keys); i++) {
         UnregisterHotKey(dwmhwnd, i);
     }
 
@@ -520,7 +520,7 @@ getroot(HWND hwnd) {
 void
 grabkeys(HWND hwnd) {
     int i;
-    for (i = 0; i < LENGTH(keys); i++) {
+    for (i = 0; i < (int)LENGTH(keys); i++) {
         RegisterHotKey(hwnd, i, keys[i].mod, keys[i].key);
     }
 }
